moved inspect_custom_regex into separate file
